Extracted HtmlRenderable
