+++ test/phlex/element_clobbering_guard.rb
@@ -5,7 +5,9 @@ end
 
 describe Phlex::HTML do
 	it "raises if you try to redefine an underscored method" do
-		expect { HTMLExample.define_method(:_h1) { nil } }.to raise_exception(Phlex::NameError)
+		expect {
+			HTMLExample.define_method(:_h1) { nil }
+		}.to raise_exception(Phlex::NameError)
 	end
 
 	it "doesn't raise if you define a new underscored method" do
